Deleting sample cromosome
