--- Samples/sample_chromosome.py
@@ -1,714 +1,0 @@
-class SampleChromosome:
-    def __init__(self):
-        self.schedule={
-            "Monday": {
-                "A": [
-                    {
-                        "teacher_id": "AP24",
-                        "subject_id": "SCS-501",
-                        "classroom_id": "R1",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "SP06",
-                        "subject_id": "TCS-503",
-                        "classroom_id": "R1",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "AK23",
-                        "subject_id": "CSP-501",
-                        "classroom_id": "R1",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "SS03",
-                        "subject_id": "TCS-502",
-                        "classroom_id": "R1",
-                        "time_slot": "12:05 - 1:00"
-                    },
-                    {
-                        "teacher_id": "BJ10",
-                        "subject_id": "TMA-502",
-                        "classroom_id": "R1",
-                        "time_slot": "1:20 - 2:15"
-                    },
-                    {
-                        "teacher_id": "SJ16",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R1",
-                        "time_slot": "2:15 - 3:10"
-                    },
-                    {
-                        "teacher_id": "AB01",
-                        "subject_id": "TCS-531",
-                        "classroom_id": "R1",
-                        "time_slot": "3:30 - 4:25"
-                    }
-                ],
-                "B": [
-                    {
-                        "teacher_id": "RS11",
-                        "subject_id": "TMA-502",
-                        "classroom_id": "R2",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "AB17",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R2",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "PM14",
-                        "subject_id": "PMA-502",
-                        "classroom_id": "L4",
-                        "time_slot": "12:05 - 1:00"
-                    },
-                    {
-                        "teacher_id": "PM14",
-                        "subject_id": "PMA-502",
-                        "classroom_id": "L4",
-                        "time_slot": "11:10 - 12:05"
-                    }
-                ],
-                "C": [
-                    {
-                        "teacher_id": "DP07",
-                        "subject_id": "TCS-503",
-                        "classroom_id": "R3",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "HP18",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R3",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "DT20",
-                        "subject_id": "XCS-501",
-                        "classroom_id": "R3",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "PK02",
-                        "subject_id": "TCS-531",
-                        "classroom_id": "R3",
-                        "time_slot": "12:05 - 1:00"
-                    }
-                ],
-                "D": [
-                    {
-                        "teacher_id": "AA04",
-                        "subject_id": "TCS-502",
-                        "classroom_id": "R4",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "AC05",
-                        "subject_id": "TCS-503",
-                        "classroom_id": "R4",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "AD08",
-                        "subject_id": "PCS-506",
-                        "classroom_id": "L2",
-                        "time_slot": "12:05 - 1:00"
-                    },
-                    {
-                        "teacher_id": "AD08",
-                        "subject_id": "PCS-506",
-                        "classroom_id": "L2",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "AB01",
-                        "subject_id": "TCS-531",
-                        "classroom_id": "R4",
-                        "time_slot": "1:20 - 2:15"
-                    },
-                    {
-                        "teacher_id": "SG19",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R4",
-                        "time_slot": "2:15 - 3:10"
-                    },
-                    {
-                        "teacher_id": "AK23",
-                        "subject_id": "CSP-501",
-                        "classroom_id": "R4",
-                        "time_slot": "3:30 - 4:25"
-                    }
-                ]
-            },
-            "Tuesday": {
-                "A": [
-                    {
-                        "teacher_id": "DT20",
-                        "subject_id": "XCS-501",
-                        "classroom_id": "R1",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "SJ16",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R1",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "RS11",
-                        "subject_id": "PCS-503",
-                        "classroom_id": "L3",
-                        "time_slot": "12:05 - 1:00"
-                    },
-                    {
-                        "teacher_id": "RS11",
-                        "subject_id": "PCS-503",
-                        "classroom_id": "L3",
-                        "time_slot": "11:10 - 12:05"
-                    }
-                ],
-                "B": [
-                    {
-                        "teacher_id": "SS03",
-                        "subject_id": "TCS-502",
-                        "classroom_id": "R2",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "AB01",
-                        "subject_id": "TCS-531",
-                        "classroom_id": "R2",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "AB17",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R2",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "AK23",
-                        "subject_id": "CSP-501",
-                        "classroom_id": "R2",
-                        "time_slot": "12:05 - 1:00"
-                    },
-                    {
-                        "teacher_id": "BJ10",
-                        "subject_id": "TMA-502",
-                        "classroom_id": "R2",
-                        "time_slot": "1:20 - 2:15"
-                    },
-                    {
-                        "teacher_id": "AP24",
-                        "subject_id": "SCS-501",
-                        "classroom_id": "R2",
-                        "time_slot": "2:15 - 3:10"
-                    },
-                    {
-                        "teacher_id": "PA21",
-                        "subject_id": "XCS-501",
-                        "classroom_id": "R2",
-                        "time_slot": "3:30 - 4:25"
-                    }
-                ],
-                "C": [
-                    {
-                        "teacher_id": "AA04",
-                        "subject_id": "TCS-502",
-                        "classroom_id": "R3",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "NB22",
-                        "subject_id": "XCS-501",
-                        "classroom_id": "R3",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "DP07",
-                        "subject_id": "PCS-503",
-                        "classroom_id": "L2",
-                        "time_slot": "12:05 - 1:00"
-                    },
-                    {
-                        "teacher_id": "DP07",
-                        "subject_id": "PCS-503",
-                        "classroom_id": "L2",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "PK02",
-                        "subject_id": "TCS-531",
-                        "classroom_id": "R3",
-                        "time_slot": "1:20 - 2:15"
-                    },
-                    {
-                        "teacher_id": "SP06",
-                        "subject_id": "TCS-503",
-                        "classroom_id": "R3",
-                        "time_slot": "2:15 - 3:10"
-                    },
-                    {
-                        "teacher_id": "HP18",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R3",
-                        "time_slot": "3:30 - 4:25"
-                    }
-                ],
-                "D": [
-                    {
-                        "teacher_id": "SG19",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R4",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "AK23",
-                        "subject_id": "CSP-501",
-                        "classroom_id": "R4",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "AP24",
-                        "subject_id": "SCS-501",
-                        "classroom_id": "R4",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "DT20",
-                        "subject_id": "XCS-501",
-                        "classroom_id": "R4",
-                        "time_slot": "12:05 - 1:00"
-                    }
-                ]
-            },
-            "Wednesday": {
-                "A": [
-                    {
-                        "teacher_id": "SJ16",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R1",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "AP24",
-                        "subject_id": "SCS-501",
-                        "classroom_id": "R1",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "RS11",
-                        "subject_id": "PCS-503",
-                        "classroom_id": "L5",
-                        "time_slot": "12:05 - 1:00"
-                    },
-                    {
-                        "teacher_id": "RS11",
-                        "subject_id": "PCS-503",
-                        "classroom_id": "L5",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "SP06",
-                        "subject_id": "TCS-503",
-                        "classroom_id": "R1",
-                        "time_slot": "1:20 - 2:15"
-                    },
-                    {
-                        "teacher_id": "BJ10",
-                        "subject_id": "TMA-502",
-                        "classroom_id": "R1",
-                        "time_slot": "2:15 - 3:10"
-                    },
-                    {
-                        "teacher_id": "AK23",
-                        "subject_id": "CSP-501",
-                        "classroom_id": "R1",
-                        "time_slot": "3:30 - 4:25"
-                    }
-                ],
-                "B": [
-                    {
-                        "teacher_id": "AB17",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R2",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "SS03",
-                        "subject_id": "TCS-502",
-                        "classroom_id": "R2",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "RS11",
-                        "subject_id": "TMA-502",
-                        "classroom_id": "R2",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "AK23",
-                        "subject_id": "CSP-501",
-                        "classroom_id": "R2",
-                        "time_slot": "12:05 - 1:00"
-                    }
-                ],
-                "C": [
-                    {
-                        "teacher_id": "AD08",
-                        "subject_id": "PCS-506",
-                        "classroom_id": "L3",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "AD08",
-                        "subject_id": "PCS-506",
-                        "classroom_id": "L3",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "DT20",
-                        "subject_id": "XCS-501",
-                        "classroom_id": "R3",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "AK23",
-                        "subject_id": "CSP-501",
-                        "classroom_id": "R3",
-                        "time_slot": "12:05 - 1:00"
-                    },
-                    {
-                        "teacher_id": "DP07",
-                        "subject_id": "TCS-503",
-                        "classroom_id": "L2",
-                        "time_slot": "2:15 - 3:10"
-                    },
-                    {
-                        "teacher_id": "DP07",
-                        "subject_id": "TCS-503",
-                        "classroom_id": "L2",
-                        "time_slot": "1:20 - 2:15"
-                    },
-                    {
-                        "teacher_id": "AB01",
-                        "subject_id": "TCS-531",
-                        "classroom_id": "R3",
-                        "time_slot": "3:30 - 4:25"
-                    }
-                ],
-                "D": [
-                    {
-                        "teacher_id": "RD09",
-                        "subject_id": "PCS-506",
-                        "classroom_id": "L1",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "RD09",
-                        "subject_id": "PCS-506",
-                        "classroom_id": "L1",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "AC05",
-                        "subject_id": "TCS-503",
-                        "classroom_id": "R4",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "JM12",
-                        "subject_id": "TMA-502",
-                        "classroom_id": "R4",
-                        "time_slot": "12:05 - 1:00"
-                    }
-                ]
-            },
-            "Thursday": {
-                "A": [
-                    {
-                        "teacher_id": "DT20",
-                        "subject_id": "XCS-501",
-                        "classroom_id": "R1",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "SP06",
-                        "subject_id": "TCS-503",
-                        "classroom_id": "R1",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "AP24",
-                        "subject_id": "SCS-501",
-                        "classroom_id": "R1",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "BJ10",
-                        "subject_id": "TMA-502",
-                        "classroom_id": "R1",
-                        "time_slot": "12:05 - 1:00"
-                    }
-                ],
-                "B": [
-                    {
-                        "teacher_id": "DP07",
-                        "subject_id": "TCS-503",
-                        "classroom_id": "R2",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "SS03",
-                        "subject_id": "TCS-502",
-                        "classroom_id": "R2",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "AB01",
-                        "subject_id": "TCS-531",
-                        "classroom_id": "R2",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "PA21",
-                        "subject_id": "XCS-501",
-                        "classroom_id": "R2",
-                        "time_slot": "12:05 - 1:00"
-                    },
-                    {
-                        "teacher_id": "RS11",
-                        "subject_id": "TMA-502",
-                        "classroom_id": "R2",
-                        "time_slot": "1:20 - 2:15"
-                    },
-                    {
-                        "teacher_id": "AP24",
-                        "subject_id": "SCS-501",
-                        "classroom_id": "R2",
-                        "time_slot": "2:15 - 3:10"
-                    },
-                    {
-                        "teacher_id": "AK23",
-                        "subject_id": "CSP-501",
-                        "classroom_id": "R2",
-                        "time_slot": "3:30 - 4:25"
-                    }
-                ],
-                "C": [
-                    {
-                        "teacher_id": "RS11",
-                        "subject_id": "PCS-503",
-                        "classroom_id": "L4",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "RS11",
-                        "subject_id": "PCS-503",
-                        "classroom_id": "L4",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "PM14",
-                        "subject_id": "PMA-502",
-                        "classroom_id": "L1",
-                        "time_slot": "12:05 - 1:00"
-                    },
-                    {
-                        "teacher_id": "PM14",
-                        "subject_id": "PMA-502",
-                        "classroom_id": "L1",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "AP24",
-                        "subject_id": "SCS-501",
-                        "classroom_id": "L5",
-                        "time_slot": "2:15 - 3:10"
-                    },
-                    {
-                        "teacher_id": "AP24",
-                        "subject_id": "SCS-501",
-                        "classroom_id": "L5",
-                        "time_slot": "1:20 - 2:15"
-                    },
-                    {
-                        "teacher_id": "SJ16",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R3",
-                        "time_slot": "3:30 - 4:25"
-                    }
-                ],
-                "D": [
-                    {
-                        "teacher_id": "AK23",
-                        "subject_id": "CSP-501",
-                        "classroom_id": "R4",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "PK02",
-                        "subject_id": "TCS-531",
-                        "classroom_id": "R4",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "JM12",
-                        "subject_id": "TMA-502",
-                        "classroom_id": "R4",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "AB17",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R4",
-                        "time_slot": "12:05 - 1:00"
-                    }
-                ]
-            },
-            "Friday": {
-                "A": [
-                    {
-                        "teacher_id": "SJ16",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R1",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "SS03",
-                        "subject_id": "TCS-502",
-                        "classroom_id": "R1",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "AP24",
-                        "subject_id": "SCS-501",
-                        "classroom_id": "R1",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "AK23",
-                        "subject_id": "CSP-501",
-                        "classroom_id": "R1",
-                        "time_slot": "12:05 - 1:00"
-                    }
-                ],
-                "B": [
-                    {
-                        "teacher_id": "BJ10",
-                        "subject_id": "TMA-502",
-                        "classroom_id": "R2",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "AK23",
-                        "subject_id": "CSP-501",
-                        "classroom_id": "R2",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "AP24",
-                        "subject_id": "SCS-501",
-                        "classroom_id": "R2",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "AA04",
-                        "subject_id": "TCS-502",
-                        "classroom_id": "R2",
-                        "time_slot": "12:05 - 1:00"
-                    },
-                    {
-                        "teacher_id": "RS11",
-                        "subject_id": "PCS-503",
-                        "classroom_id": "L3",
-                        "time_slot": "2:15 - 3:10"
-                    },
-                    {
-                        "teacher_id": "RS11",
-                        "subject_id": "PCS-503",
-                        "classroom_id": "L3",
-                        "time_slot": "1:20 - 2:15"
-                    },
-                    {
-                        "teacher_id": "AB01",
-                        "subject_id": "TCS-531",
-                        "classroom_id": "R2",
-                        "time_slot": "3:30 - 4:25"
-                    }
-                ],
-                "C": [
-                    {
-                        "teacher_id": "RS11",
-                        "subject_id": "TMA-502",
-                        "classroom_id": "R3",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "AC05",
-                        "subject_id": "TCS-502",
-                        "classroom_id": "R3",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "AP24",
-                        "subject_id": "SCS-501",
-                        "classroom_id": "R3",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "SP06",
-                        "subject_id": "TCS-503",
-                        "classroom_id": "R3",
-                        "time_slot": "12:05 - 1:00"
-                    },
-                    {
-                        "teacher_id": "PM14",
-                        "subject_id": "PMA-502",
-                        "classroom_id": "L5",
-                        "time_slot": "2:15 - 3:10"
-                    },
-                    {
-                        "teacher_id": "PM14",
-                        "subject_id": "PMA-502",
-                        "classroom_id": "L5",
-                        "time_slot": "1:20 - 2:15"
-                    },
-                    {
-                        "teacher_id": "AB17",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R3",
-                        "time_slot": "3:30 - 4:25"
-                    }
-                ],
-                "D": [
-                    {
-                        "teacher_id": "SS03",
-                        "subject_id": "TCS-502",
-                        "classroom_id": "R4",
-                        "time_slot": "9:00 - 9:55"
-                    },
-                    {
-                        "teacher_id": "AK23",
-                        "subject_id": "CSP-501",
-                        "classroom_id": "R4",
-                        "time_slot": "9:55 - 10:50"
-                    },
-                    {
-                        "teacher_id": "AP24",
-                        "subject_id": "SCS-501",
-                        "classroom_id": "R4",
-                        "time_slot": "11:10 - 12:05"
-                    },
-                    {
-                        "teacher_id": "HP18",
-                        "subject_id": "TCS-509",
-                        "classroom_id": "R4",
-                        "time_slot": "12:05 - 1:00"
-                    }
-                ]
-            }
-        }
